+++ app/colors/index.js
