+++ app/components/MovieCard/index.js
