+++ app/components/Wizard/styles.js
