--- app/constants/Colors.js
@@ -1,7 +1,0 @@
-export default {
-  blackColor: '#23222E',
-  redColor: '#FF5656',
-  grayColor: '#EDF2F6',
-  blueColor: '#6A7EFC',
-  darkBlackColor: '#212121'
-};
