--- app/constants/poster_path.js
@@ -1,2 +1,0 @@
-export const POSTER = 'https://image.tmdb.org/t/p/w500';
-export const PATH_POPULAR = 'https://movied.herokuapp.com/popular';
