+++ app/screens/DiscoverScreen/components/Card.js
