+++ app/screens/DiscoverScreen/components/SurveyNav.js
