+++ app/screens/DiscoverScreen/components/index.js
