+++ app/screens/DiscoverScreen/components/styles/Card.js
