+++ app/screens/DiscoverScreen/styles.js
