+++ app/screens/LikedListScreen/components/MovieItem.js
