+++ app/screens/LoginScreen/spotifyIconBlack.png
