+++ app/screens/RecommendedScreen/components/Card.js
