+++ app/screens/RecommendedScreen/styles.js
